Documentation automation script for Gunrock
Please run this script in 'doc' folder, inside master or dev branch.
It will automatically update PROJECT_NUMBER in gunrock.doxygen,
run doxygen and check if there are warnings, put everything
inside gh-pages in the right order and then commit.
It will not push the new commit to the upstream; you still need to
manually do that after verifying everything is good.