Use ccbin option for gencode auto-detection
When CMake runs the AutoDetectCudaArch file and executes the command
to run our test program, we don't pass in the -ccbin option. If the
user has multiple versions of gcc installed, then the default will
be used instead of the one specified to CMake via $CC or
CMAKE_C_COMPILER.

This change modifes AutoDetectCudaArch to include the
CMAKE_C_COMPILER in the test command that is executed during build.