Modify device intrinsics to compile on all arches
When compiling for compute capabilites before 3.5, the atomicMin
and __ldg device intrinsics do not exist. When compiling for these
arches, we did not have definitions for atomicMin properly defined.
This caused compilation issues when autodetection of the proper arch
failed.

This change also modfies the intersection kernel to use our internal
function _ldg instead of the device intrinsic __ldg.