Update mgpu Scan and SortedSearch with launch_box
Specifies the launch_box_t template struct for any use of moderngpu's
functions. In moderngpu 2.0 the library user specifies a launch_box
that chooses values for the number of threads and the grain size.

They are specified as mgpu::arch_$cc$_cta<$nt$, vt>, where $cc$ is the
sm version, $nt$ is the number of threads and $vt$ is the grain size
or number of values per thread. Sean Baxter recommends an odd value
for $vt$. There are further options explained in moderngpu's
documentation.

These are placeholder values for now. I put them in and saw no
issues running basic tests but we should test these values in more
detail and consider finding the most optimal options.