JSON writer improvements
Moves the JSON interface from using the Writer / SAX version of
RapipJSON's interface to the Document version. This allows sorting of
the document and throwing assertions for invalid JSON. Previously we
could write incomplete files when a job was interrupted. Now
incomplete jobs will have an elapsed time of -1, but otherwise be
valid JSON.

Similarly some keys could have values of NaN or INF, which caused
the previous interface to write empty string for their value. This is
because RapidJSON disallows NaN and INF (the JSON spec doesn't include
them, but Javascript and other interfaces are more permissive). Now
these invalid values will be written as null to the resulting
document.