Link with C++ compiler for shared libs example
The shared_libs examples need to be linked with a C++ compiler since
they link against a C++ library. Previously, cmake properly used g++ as
expected. It's not clear why, but CMake decided to start using gcc for
these files because they end in ".c". This may be due to a change to one
of the parent CMakeLists.txt files, but it's difficult to find why this
behavior changed.

This change explicitly tells CMake to use the C++ linker with the
set_target_properties command.