Move #ifdef checking for whole grid operations
My previous commit broke when SM = 7.0. I think that when compiling both
host and device side, the host side version no longer had the kernels
defined but the device side did.

The fix is to keep the kernel defined in all compilations but #ifdef out
the use of this_grid.