Fix mgpu memory issues from seperate stream (#706)
After updating moderngpu from 1.0 -> 2.0 we forgot to make sure that the
mgpu calls were on the same stream as our other kernels. Operators using
mgpu's SortedSearch and Scan functions expected them to be complete
before the next kernel call, but since the mgpu context was on the
default stream it executed in parallel with the other kernels. This
caused the output_offsets array used in the following kernel to be incorrect,
and out of bounds memory accesses would occur.

This fix ensures that the mgpu::standard_context_t is setup to use the
same stream as the other kernels. To do so, we keep a context pointer,
allocate it on the heap when the operator parameters are intialized, and
free it when they are released.