CXX Interface for Subgraph Matching and HITS
This change adds the file gunrock.hpp to serve as a CXX interface to
Gunrock. The sm_template and hits_template functions are renamed as
simply sm and hits, and then exposed to the library user in gunrock.hpp.
These functions already have template instantiations from the C library
functions, so we don't need to include their implementation inside of
gunrock.hpp, and it can be a relatively small header that we distribute.

I also moved all CXX interface tests to a new folder under shared_libs
and separated them from the C interface tests. Now we can test both
versions of the interface from these folders.